requirements.txt
Django==1.6.2
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.2
static==1.0.2
wsgiref==0.1.2
django-userena==1.3.0
